pyproject.toml
[project]
name = "ticket-manager"
version = "1.0.0"
description = "Ticket manager system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "djangorestframework==3.16.0",
    "drf-yasg==1.21.10",
    "psycopg2-binary>=2.9.10",
    "requests==2.32.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-debug-toolbar==5.2.0",
    "flake8>=7.2.0",
    "ipdb==0.13.13",
    "ipython==9.3.0",
    "pytest==8.4.0",
    "pytest-bdd>=8.1.0",
    "pytest-cov>=6.2.1",
    "pytest-django==4.11.1",
]
